pyproject.toml
[tool.poetry]
name = "resdex-agent"
version = "1.0.0"
description = "ResDex Agent - AI-powered candidate search and filtering system following Google ADK patterns"
authors = ["ResDex Team <team@resdex.com>"]
readme = "README.md"
packages = [{include = "resdex_agent"}]

[tool.poetry.dependencies]
python = "^3.9"
streamlit = "^1.32.0"
pandas = "^2.0.0"
requests = "^2.31.0"
pymysql = "^1.1.0"
sqlalchemy = "^2.0.0"
pydantic = "^2.6.0"
python-dotenv = "^1.0.0"
openai = "^1.0.0"
aiohttp = "^3.8.0"
asyncio = {version = "*", markers = "python_version < '3.7'"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.0.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
coverage = "^7.4.0"

[tool.poetry.group.eval.dependencies]
pytest-benchmark = "^4.0.0"
memory-profiler = "^0.61.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true